+++ ostap/fitting/morphing_pdf.py
@@ -16,6 +16,7 @@ __all__     = (
     ##
     'MorphingN1_pdf'  , ## 1D-morphing PDF
     'MorphingN2_pdf'  , ## 2D-morphing PDF
+    'MorphingN3_pdf'  , ## 3D-morphing PDF
     'LinearMorph_pdf' , ## linear morphing 
     ##
     )
@@ -179,8 +180,7 @@ class MorphingN1_pdf (PDF1) :
         self.__mu = self.make_var (
             morph_var                ,
             "mu_%s"           % name ,
-            "morphing mu(%s)" % name ,
-            False , mu_min , mu_max )
+            "morphing mu(%s)" % name , False , mu_min , mu_max )
 
         ## vector of morphing values 
         muvct  = ROOT.TVectorD ( len ( self.pdflist ) )
@@ -230,7 +230,6 @@ class MorphingN1_pdf (PDF1) :
         return self.__setting 
 
 
-
 # =============================================================================
 ## @class MorphingN2_pdf
 #  Wrapper for <code>RooMomentMorph</code> PDF
@@ -261,7 +260,7 @@ class MorphingN2_pdf (PDF1) :
                    morph_var2 = None   , ## morphing variable mu2 
                    xvar       = None ) : ## observable (1D) 
 
-        assert pdfs and 2 <= len ( pdfs ) , \
+        assert pdfs and 4 <= len ( pdfs ) , \
                "Invalid dictionary of morphing PDFs!"
         
         if setting is None : setting = ROOT.RooMomentMorphND.Linear 
@@ -300,13 +299,13 @@ class MorphingN2_pdf (PDF1) :
         self.__mu1 = self.make_var (
             morph_var1                ,
             "mu1_%s"           % name ,
-            "morphing mu1(%s)" % name , morph_var1 , v1ps[0] , v1ps[-1] )
+            "morphing mu1(%s)" % name , False , v1ps[0] , v1ps[-1] )
         
         ## create morphing variables 
         self.__mu2 = self.make_var (
             morph_var2                ,
             "mu2_%s"           % name ,
-            "morphing mu2(%s)" % name , morph_var2 , v2ps[0] , v2ps[-1] )
+            "morphing mu2(%s)" % name , False , v2ps[0] , v2ps[-1] )
         
         self.__pdfdict = {}
         for k in sorted ( pdfs.keys()  ) :
@@ -321,7 +320,7 @@ class MorphingN2_pdf (PDF1) :
                 raise TypeError( "Invalid component type: %s/%s" % ( pdfk , type ( pdfk ) ) ) 
 
             pair = k , pdfk 
-            self.pdfdict[ ( v1, v2 ) ] = pdfk
+            self.pdfdict [ ( v1, v2 ) ] = pdfk
             
         ## save setting 
         self.__setting = setting
@@ -340,8 +339,8 @@ class MorphingN2_pdf (PDF1) :
             
             v1 , v2 = k
 
-            assert v1 in v1ps , 'Morphing2D_pdf: Invalid v1 value %s' % v1 
-            assert v2 in v2ps , 'Morphing2D_pdf: Invalid v2 value %s' % v2 
+            assert v1 in v1ps , 'MorphingN2_pdf: Invalid v1 value %s' % v1 
+            assert v2 in v2ps , 'MorphingN2_pdf: Invalid v2 value %s' % v2 
             
             ib1 = v1ps.index ( v1 ) 
             ib2 = v2ps.index ( v2 )
@@ -382,11 +381,7 @@ class MorphingN2_pdf (PDF1) :
         return self.__mu1
     @mu1.setter
     def mu1 ( self , value ) :
-        v = float ( value ) 
-        mm = self.__mu1.minmax()
-        if mm and not mm[0] <= v <= mm[1] :
-            self.error ( "Morphing parameter %s is outside [%s,%s]" % ( v , mm[0] , mm[1] ) )
-        self.__mu1.setVal ( v )
+        self.set_value  ( self.__mu1 , value )
 
     @property
     def mu2 ( self ) :
@@ -394,15 +389,11 @@ class MorphingN2_pdf (PDF1) :
         return self.__mu2
     @mu2.setter
     def mu2 ( self , value ) :
-        v = float ( value ) 
-        mm = self.__mu2.minmax()
-        if mm and not mm[0] <= v <= mm[1] :
-            self.error ( "Morphing parameter %s is outside [%s,%s]" % ( v , mm[0] , mm[1] ) )
-        self.__mu2.setVal ( v )
+        self.set_value  ( self.__mu2 , value )
 
     @property
     def grid    ( self ) :
-        """'grid' : morphing grid"""
+        """'grid' : morphing 2D-grid"""
         return self.__grid
     
     @property
@@ -415,8 +406,208 @@ class MorphingN2_pdf (PDF1) :
         """'setting': morphing setting"""
         return self.__setting 
                        
-    
+
+
+# =============================================================================
+## @class MorphingN3_pdf
+#  Wrapper for <code>RooMomentMorph</code> PDF
+#  - 1D morphing/3D  PDF   
+#  @see RooMomentMorphND
+#  @see Baak, M., Gadatsch, S., Harrington, R., & Verkerke, W. (2015).
+#       "Interpolation between multi-dimensional histograms using
+#       a new non-linear moment morphing method".
+#       Nuclear Instruments & Methods in Physics Research.
+#       Section A - Accelerators Spectrometers Detectors and Associated Equipment, 771, 39-48.
+#  @see https://doi.org/10.1016/j.nima.2014.10.033
+class MorphingN3_pdf (PDF1) :
+    """ Wrapper for ROOT.RooMomentMorphND PDF for N = 2 
+    - 1D morphing/3D PDF
+    - see ROOT.RooMomentMorphND
+    - see Baak, M., Gadatsch, S., Harrington, R., & Verkerke, W. (2015).
+    'Interpolation between multi-dimensional histograms using
+    a new non-linear moment morphing method'.
+    Nuclear Instruments & Methods in Physics Research.
+    Section A - Accelerators Spectrometers Detectors and Associated Equipment, 771, 39-48.
+    - see https://doi.org/10.1016/j.nima.2014.10.033
+    """
+    def __init__ ( self                ,
+                   name                , ## PDF name 
+                   pdfs                , ## dictionary {mu1,mu2 -> pdf }
+                   setting    = None   , ## morphing setting 
+                   morph_var1 = None   , ## morphing variable mu1 
+                   morph_var2 = None   , ## morphing variable mu2 
+                   morph_var3 = None   , ## morphing variable mu3
+                   xvar       = None ) : ## observable (1D) 
+
+        assert pdfs and 6 <= len ( pdfs ) , \
+               "Invalid dictionary of morphing PDFs!"
         
+        if setting is None : setting = ROOT.RooMomentMorphND.Linear 
+
+        assert isinstance ( setting , integer_types ) and 0 <= setting < 5,\
+               'Invalid value for the setting %s/%s' %  ( setting , type ( setting ) )
+
+        v1ps = set ()
+        v2ps = set () 
+        v3ps = set () 
+        
+        for k in pdfs :
+            v1 , v2 , v3 = k
+            v1ps.add ( v1 )
+            v2ps.add ( v2 )
+            v3ps.add ( v3 )
+            p = pdfs [ k ]
+            if not xvar and isinstance ( p , PDF1 ) :
+                xvar = p.xvar
+                
+        assert xvar and isinstance ( xvar , ROOT.RooAbsReal ) , 'Cannot deduce xvar!'
+
+        assert 2 <= len ( v1ps ) and 2 <= len ( v2ps ) and 2 <= len ( v3ps ) , 'Invalid number of bins!'
+
+        v1ps = list ( v1ps ) 
+        v2ps = list ( v2ps )
+        v3ps = list ( v3ps )
+
+        v1ps.sort ()
+        v2ps.sort ()
+        v3ps.sort ()
+
+        assert  len ( pdfs ) ==  len ( v1ps ) * len ( v2ps ) * len ( v3ps ) ,\
+               'Invalid table/dict structure!'
+        
+        ## initialize the base class 
+        PDF1.__init__ ( self , name , xvar )
+
+        ## create morphing variables 
+        self.__mu1 = self.make_var (
+            morph_var1                ,
+            "mu1_%s"           % name ,
+            "morphing mu1(%s)" % name , False , v1ps[0] , v1ps[-1] )
+        
+        ## create morphing variables 
+        self.__mu2 = self.make_var (
+            morph_var2                ,
+            "mu2_%s"           % name ,
+            "morphing mu2(%s)" % name , False , v2ps[0] , v2ps[-1] )
+        
+        ## create morphing variables 
+        self.__mu3 = self.make_var (
+            morph_var3                ,
+            "mu3_%s"           % name ,
+            "morphing mu3(%s)" % name , False , v3ps[0] , v3ps[-1] )
+        
+        self.__pdfdict = {}
+        for k in sorted ( pdfs.keys()  ) :
+
+            v1 , v2 . v3 = k
+            
+            pdfk = pdfs [ k ] 
+            if   isinstance ( pdfk , PDF1           ) and pdfk.xvar is self.xvar : pass 
+            elif isinstance ( pdfk , ROOT.RooAbsPdf ) : 
+                pdfk = Generic1D_pdf ( pdfk , xvar = self.xvar ) 
+            else :
+                raise TypeError( "Invalid component type: %s/%s" % ( pdfk , type ( pdfk ) ) ) 
+
+            pair = k , pdfk 
+            self.pdfdict [ ( v1, v2 , v3 ) ] = pdfk
+            
+        ## save setting 
+        self.__setting = setting
+
+        
+        ## fill morphing grid
+        from ostap.fitting.variables import binning
+        
+        bins_v1     = binning ( v1ps , name = 'morph1' ) 
+        bins_v2     = binning ( v2ps , name = 'morph2' ) 
+        bins_v3     = binning ( v3ps , name = 'morph3' ) 
+        self.__grid = ROOT.RooMomentMorphND.Grid ( bins_v1 , bins_v2 . bins_v3 ) 
+        
+        for k in self.pdfdict :
+
+            p       = self.pdfdict [ k ]
+            
+            v1 , v2 , v3 = k
+            
+            assert v1 in v1ps , 'MorphingN3_pdf: Invalid v1 value %s' % v1 
+            assert v2 in v2ps , 'MorphingN3_pdf: Invalid v2 value %s' % v2 
+            assert v3 in v3ps , 'MorphingN3_pdf: Invalid v3 value %s' % v3
+            
+            ib1 = v1ps.index ( v1 ) 
+            ib2 = v2ps.index ( v2 )
+            ib3 = v3ps.index ( v3 )
+            
+            ## ib1 = bins_v1.binNumber ( v1 )
+            ## ib2 = bins_v2.binNumber ( v2 )
+
+            self.__grid.addPdf ( p.pdf , ib1 , ib2 , ib3 )
+
+        morph_vars  = ROOT.RooArgList ( self.mu1 , self.mu2 , self.mu3 )
+        observables = ROOT.RooArgList ( self.xvar )
+
+        self.aux_keep.append  ( morph_vars  )
+        self.aux_keep.append  ( observables )
+        
+        ## create the PDF  
+        self.pdf = ROOT.RooMomentMorphND (
+            self.roo_name ( 'morph3_' )   ,
+            "Morphing 3D %s" % self.name  , 
+            morph_vars              , ## morphing variables 
+            observables             , ## observables 
+            self.grid               , ## morphing grid 
+            self.setting            ) ## morphing setting 
+
+        #
+        self.config = {
+            'name'       : self.name    ,
+            'setting'    : self.setting ,
+            'xvar'       : self.xvar    ,
+            'morph_var1' : self.mu1     ,
+            'morph_var2' : self.mu2     ,
+            'morph_var3' : self.mu3     ,
+            'pdfs'       : self.pdfdict ,
+            }
+        
+    @property
+    def mu1 ( self ) :
+        """'mu1' : the first morphing variable"""
+        return self.__mu1
+    @mu1.setter
+    def mu1 ( self , value ) :
+        self.set_value  ( self.__mu1 , value )
+
+    @property
+    def mu2 ( self ) :
+        """'mu2' : the second morphing variable"""
+        return self.__mu2
+    @mu2.setter
+    def mu2 ( self , value ) :
+        self.set_value  ( self.__mu2 , value )
+
+    @property
+    def mu3 ( self ) :
+        """'mu3' : the third morphing variable"""
+        return self.__mu3
+    @mu3.setter
+    def mu3 ( self , value ) :
+        self.set_value  ( self.__mu3 , value )
+
+    @property
+    def grid    ( self ) :
+        """'grid' : morphing 2D-grid"""
+        return self.__grid
+    
+    @property
+    def pdfdict ( self ) :
+        """'pdfdict' : Dictionary { morphing parameters : pdf } """
+        return self.__pdfdict
+
+    @property 
+    def setting ( self ) :
+        """'setting': morphing setting"""
+        return self.__setting 
+                       
+
 # =============================================================================
 if '__main__' == __name__ :